Move all types to a dedicated types.ts file (as we always do)
