+++ packages/statemanager/src/index.ts
@@ -4,3 +4,4 @@ export * from './rpcStateManager.js'
 export * from './simpleStateManager.js'
 export * from './statelessVerkleStateManager.js'
 export * from './stateManager.js'
+export * from './types.js'
